+++ examples/example_02_respiration.py
@@ -1,6 +1,6 @@
 '''
-Respiration example
-===================
+Respiration tutorial
+====================
 
 
 '''
@@ -12,20 +12,12 @@ from pprint import pprint
 
 import physio
 
-##############################################################################
-# 
-# physio.respiration
-# 
-# ------------------
-#
-# 
-#  
+
 
 ##############################################################################
 # 
-# Respiration : quick way
-# 
-# -----------------------
+# Respiration cucle detection: quick way
+# --------------------------------------
 #
 #  The fastest way is to use compute_respiration() using predefine parameters set
 #  here a simple example
@@ -37,7 +29,7 @@ srate = 1000.
 times = np.arange(raw_resp.size) / srate
 
 # the easiest way is to use predefined parameters
-resp, resp_cycles = physio.compute_respiration(raw_resp, srate, parameter_set='human_airflow')
+resp, resp_cycles = physio.compute_respiration(raw_resp, srate, parameter_preset='human_airflow')
 
 # resp_cycles is a dataframe containing all cycles and related fetaures (duration, amplitude, volume, timing)
 print(resp_cycles)
@@ -57,8 +49,8 @@ ax.set_xlim(110, 170)
 
 ##############################################################################
 # 
-# Parameters tuning
-# -----------------------
+# Cycle detection : Parameters tuning
+# -----------------------------------
 # 
 # Here a simple recipe to change some predefined parameters
 # We change here the length smoothing parameter