+++ doc/release_notes.rst
@@ -10,13 +10,16 @@ Release notes
 Version 0.1.0
 =============
 
+2023-06-XX
+
 **Initial Version with**:
 
   * respiration cycle detection
-  * respiration cycle features (amplitude, duration, volumns ...)
+  * respiration cycle features (amplitude, duration, volumes ...)
   * simple preprocess on signal : filter using scipy and smoothing
   * ecg peak detection
-  * ecg metrics
+  * ecg/hrv metrics (time domain and freq domain)
+  * rsa : new approach to get cycle-per-cycle metrics
   * cyclic deformation machinery : a simple strecher of any signal to cycle template
   * simple reader of micromed and brainvision using neo
-
+  * "automagic" parameters for differents species